refactor: move `CInstantSendLock` definition to `instantsend/lock.{cpp,h}`
Review with `git log -p -n1 --color-moved=dimmed_zebra`.
